+++ atrium-api/src/app/bsky/feed/search_posts.rs
@@ -7,7 +7,7 @@ pub struct Parameters {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub cursor: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub limit: Option<i32>,
+    pub limit: Option<crate::types::LimitedNonZeroU8<100u8>>,
     ///Search query string; syntax, phrase, boolean, and faceting is unspecified, but Lucene query syntax is recommended.
     pub q: String,
 }
@@ -18,7 +18,7 @@ pub struct Output {
     pub cursor: Option<String>,
     ///Count of search hits. Optional, may be rounded/truncated, and may not be possible to paginate through all hits.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub hits_total: Option<i32>,
+    pub hits_total: Option<i64>,
     pub posts: Vec<crate::app::bsky::feed::defs::PostView>,
 }
 #[derive(serde::Serialize, serde::Deserialize, Debug, Clone, PartialEq, Eq)]